% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PODES.R
\name{PODES}
\alias{PODES}
\title{Create Village Potential Statistics of Indonesia}
\usage{
PODES(data)
}
\arguments{
\item{data}{dataframes}
}
\value{
A dataframes of Village Potential Statistics of Indonesia
}
\description{
Create Village Potential Statistics of Indonesia
}
\examples{
PODES(PODES2019)

}
